requirements.txt
aiomysql==0.2.0
anyio==4.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
databases==0.9.0
fastapi==0.95.2
h11==0.14.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
numpy==1.26.4
pandas==2.0.1
pydantic==1.10.7
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.27.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.22.0